+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/), and this 
 
 ## [Release candidate] - started 2025-01
 
+## [3.1.1] - 2025-01-15
+
+### Fixed
+
+- ValidationError for IntersectResponse in `/api/v0/modification/target/RBP?` (Bed6Record DTO).
+- EUFID link in ModificationSiteTable.
+
+### Changed
+
+- dev and tests dependencies
+- node/npm version
+
 ## [3.1.0] - 2025-01-08
 
 ### Added